feat(data): move db.rs to db/mod.rs and db/tests.rs
